fix maximizing windows then switching windows would kill the maximized windows

+++ main.cpp
@@ -4,7 +4,82 @@
 
 const int defaultDelay = 1;
 
-unsigned int *getActiveWindowsPID()
+class Vector2D
+{
+public:
+    int width;
+    int height;
+    int x;
+    int y;
+};
+
+Vector2D getWindowCoordinates(unsigned int pid)
+{
+    CFArrayRef windows = CGWindowListCopyWindowInfo(kCGWindowListOptionAll, kCGNullWindowID);
+    for (CFIndex i = 0; i < CFArrayGetCount(windows); i++)
+    {
+        CFDictionaryRef windowInfo = (CFDictionaryRef)CFArrayGetValueAtIndex(windows, i);
+        CFNumberRef pidNumber = (CFNumberRef)CFDictionaryGetValue(windowInfo, kCGWindowOwnerPID);
+        unsigned int windowPid;
+        if (CFNumberGetValue(pidNumber, kCFNumberSInt32Type, &windowPid))
+        {
+            if (windowPid == pid)
+            {
+                Vector2D vector;
+                CGRect bounds;
+                CFDictionaryRef boundsRef = (CFDictionaryRef)CFDictionaryGetValue(windowInfo, kCGWindowBounds);
+                CGRectMakeWithDictionaryRepresentation(boundsRef, &bounds);
+                vector.width = bounds.size.width;
+                vector.height = bounds.size.height;
+                vector.x = bounds.origin.x;
+                vector.y = bounds.origin.y;
+                return vector;
+            }
+        }
+    }
+    return Vector2D();
+}
+
+Vector2D getMonitorResolution(unsigned int pid)
+{
+    CFArrayRef windows = CGWindowListCopyWindowInfo(kCGWindowListOptionAll, kCGNullWindowID);
+
+    for (CFIndex i = 0; i < CFArrayGetCount(windows); i++)
+    {
+        CFDictionaryRef windowInfo = (CFDictionaryRef)CFArrayGetValueAtIndex(windows, i);
+        CFNumberRef ownerPIDRef = (CFNumberRef)CFDictionaryGetValue(windowInfo, kCGWindowOwnerPID);
+
+        pid_t ownerPID;
+        if (CFNumberGetValue(ownerPIDRef, kCFNumberSInt32Type, &ownerPID) && ownerPID == pid)
+        {
+            CGRect windowBounds;
+            CFDictionaryRef boundsRef = (CFDictionaryRef)CFDictionaryGetValue(windowInfo, kCGWindowBounds);
+            CGRectMakeWithDictionaryRepresentation(boundsRef, &windowBounds);
+
+            int monitorNumber = 0;
+            CGDirectDisplayID displayIDs[10];
+            uint32_t displayCount;
+            CGGetDisplaysWithPoint(windowBounds.origin, 10, displayIDs, &displayCount);
+
+            for (uint32_t j = 0; j < displayCount; j++)
+            {
+                if (CGDisplayIsActive(displayIDs[j]))
+                {
+                    CGRect displayBounds = CGDisplayBounds(displayIDs[j]);
+                    Vector2D vector;
+                    vector.width = displayBounds.size.width;
+                    vector.height = displayBounds.size.height;
+                    return vector;
+                }
+            }
+        }
+    }
+
+    CFRelease(windows);
+    return Vector2D();
+}
+
+unsigned int *getActiveWindows()
 {
     CFArrayRef windows = CGWindowListCopyWindowInfo(kCGWindowListOptionOnScreenBelowWindow, kCGNullWindowID);
     unsigned int *pids = new unsigned int[CFArrayGetCount(windows)];
@@ -44,11 +119,11 @@ bool isWindowMinimized(unsigned int pid)
 
 int main()
 {
-    unsigned int *previous = getActiveWindowsPID();
+    unsigned int *previous = getActiveWindows();
     while (1)
     {
         sleep(defaultDelay);
-        unsigned int *current = getActiveWindowsPID();
+        unsigned int *current = getActiveWindows();
         for (int i = 1; i < previous[0]; ++i)
         {
             bool found = false;
@@ -60,8 +135,8 @@ int main()
                     break;
                 }
             }
-            
-            if (!found && !isWindowMinimized(previous[i]))
+
+            if (!found && !isWindowMinimized(previous[i]) && getWindowCoordinates(previous[i]).width != getMonitorResolution(previous[i]).width)
             {
                 kill(previous[i], SIGTERM);
             }
